backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.1.0